+++ nodalia-cards.manifest.js
@@ -1,11 +1,11 @@
 export default {
-  "pkgVersion": "2.2.4",
-  "contentSha256_12": "372f814c0f18",
+  "pkgVersion": "2.2.5",
+  "contentSha256_12": "201d7308ec5d",
   "file": "nodalia-cards.js",
   "loaderFile": "nodalia-cards.js",
   "hacsFile": "nodalia-cards.js",
   "editorSha256_12": "9722b9c0fc8f"
 };
-export const pkgVersion = "2.2.4";
-export const contentSha256_12 = "372f814c0f18";
+export const pkgVersion = "2.2.5";
+export const contentSha256_12 = "201d7308ec5d";
 export const file = "nodalia-cards.js";